# slices: improve the performance of Equal