Although it does an additional index operation, but it avoids the value copy of the range loop,
and it will has a better performance if E is a large structure.

Compares two int slices, the slices length are 10:
name                old time/op  new time/op  delta
Equal_WithIntSlices  5.80ns ± 0%  5.80ns ± 1%   ~     (p=0.889 n=14+15)

Compares two string slices, the slices length are 10:
name                   old time/op  new time/op  delta
Equal_WithStringSlices  17.6ns ± 1%  16.6ns ± 0%  -5.61%  (p=0.000 n=15+15)

Compares two runtime.MemStats slices, the slices length are 10:
name                        old time/op  new time/op  delta
Equal_WithLagerStructSlices  4.46µs ± 0%  1.53µs ± 1%  -65.67%  (p=0.000 n=15+15)